requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Telegram Bot
python-telegram-bot==20.7

# Instagram Download
yt-dlp==2024.1.16

# Local Whisper (faster-whisper)
faster-whisper==1.0.0

# Ollama Local LLM
ollama==0.4.0

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# Scheduler
apscheduler==3.10.4

# HTTP Client
httpx==0.26.0
aiofiles==23.2.1

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
